:tada: feat: add `callsign2` to combined units
