+++ packages/api/prisma/schema.prisma
@@ -753,6 +753,7 @@ model CombinedLeoUnit {
   id             String         @id @default(uuid())
   officers       Officer[]
   callsign       String
+  callsign2      String?
   radioChannelId String?
   status         StatusValue?   @relation("combinedUnitStatusToValue", fields: [statusId], references: [id])
   statusId       String?
